Update: index and testing file now linked
The index.js file now serves as the hub for all of the modules that
will need to be created during the project. The testing file now points
to index.js and watches it so I can get some real time feedback. I have
also written an empty test just to confirm that the testing link works
on it's own.